亚马逊补货/requirements.txt
# 领星ERP补货数据获取工具依赖包

# HTTP请求库
requests>=2.28.0
urllib3>=1.26.0

# 数据处理
pandas>=1.5.0
numpy>=1.21.0

# Excel文件处理
openpyxl>=3.0.10

# 加密库
pycryptodome>=3.15.0

# 日期时间处理
python-dateutil>=2.8.0

# 数据库连接器
# MySQL支持（可选）
# PyMySQL>=1.0.0
# mysqlclient>=2.1.0

# PostgreSQL支持（可选）
# psycopg2-binary>=2.9.0

# SQLite支持（Python内置，无需安装）

# 服务器相关
Flask>=2.3.0  # Web服务器框架（飞书webhook必需）
# gunicorn>=20.1.0  # 生产环境WSGI服务器（Linux）
python-dotenv>=1.0.0  # 环境变量加载（配置文件支持）

# 系统监控
psutil>=5.9.0  # 系统资源监控

# Windows服务支持
pywin32>=306  # Windows服务支持（仅Windows系统需要）

# JSON处理（Python内置，无需安装）
# json

# 文件路径处理（Python内置，无需安装）
# os
# sys

# 正则表达式（Python内置，无需安装）
# re

# 时间处理（Python内置，无需安装）
# time
# datetime

# 数据类（Python 3.7+内置，无需安装）
# dataclasses

# 类型提示（Python 3.5+内置，无需安装）
# typing

# 日志记录（Python内置，无需安装）
# logging

# 命令行参数解析（Python内置，无需安装）
# argparse

# Base64编码（Python内置，无需安装）
# base64

# URL编码（Python内置，无需安装）
# urllib.parse

# MD5哈希（Python内置，无需安装）
# hashlib